LDBR-1.16: Исправить страницу входа, исправить код под новый стиль.
