+++ src/pages/LoginPage/LoginPage.hbs
@@ -1,4 +1,4 @@
-<link rel="stylesheet" href="/src/pages/login/login.css">
+<link rel="stylesheet" href="/src/pages/LoginPage/LoginPage.css">
 <link rel="stylesheet" href="../../../public/css/popups.css">
 
 <header id="header-main"></header>
