--- src/utils/constants.js
@@ -1,13 +1,0 @@
-'use strict';
-
-/**
- * Константа, содержащая в себе urlы.
- */
-export const urls = {
-    root: '/',
-    boards: '/boards',
-    login: '/login',
-    register: '/register',
-    logout: '/logout',
-    404: '404',
-};
